[TASK-0349] Сделал StoryListPage и API для получения списка историй (#40)
* [TASK-0349] Сделал StoryListPage и API для получения списка историй

* [TASK-0349] Вывод карточек истории, но пока без корректной сквозной нумерации

* [TASK-0349] При выводе истории пересоздаю и упорядочиваю карточки

* [TASK-0349] В историю добавлен текущий фрагмент

* [TASK-0349] Добавлен сброс кэша историй при её изменениях

* [TASK-0349] Убрано дублирование в конструкторах StoryCard

---------

+++ src/Domain/YagoEntities/YagoEntity.cs
@@ -4,12 +4,12 @@ namespace YAGO.World.Domain.YagoEntities
 {
     public class YagoEntity
     {
-        public int Id { get; }
+        public long Id { get; }
         public YagoEntityType Type { get; }
         public string Name { get; }
 
         public YagoEntity(
-            int id,
+            long id,
             YagoEntityType type,
             string name)
         {
